Moved GENN-unrelated UI functions to helper class
